% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-vector.R
\name{as.vector.torch_tensor}
\alias{as.vector.torch_tensor}
\title{Convert \code{torch_tensor} to a vector}
\usage{
\method{as.vector}{torch_tensor}(x, mode = "any")
}
\description{
Convert \code{torch_tensor} to a vector
}
