% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-ts-dataset.R
\name{as_ts_dataset}
\alias{as_ts_dataset}
\title{Create a torch dataset for time series data}
\usage{
as_ts_dataset(
  data,
  formula,
  index = NULL,
  key = NULL,
  target = NULL,
  n_timesteps,
  h = 1,
  sample_frac = 1
)
}
\arguments{
\item{data}{(data.frame)}

\item{formula}{A formula describing, how to use the data}

\item{index}{The index column}

\item{key}{The key column(s)}

\item{n_timesteps}{The time seris chunk length}

\item{h}{Forecast horizon}

\item{sample_frac}{Sample a fraction of rows (default: 1, i.e.: all the rows)}
}
\description{
Create a torch dataset for time series data
}
\examples{
data_set <-
 read.csv("https://raw.githubusercontent.com/jbrownlee/Datasets/master/daily-min-temperatures.csv")

data_split <- initial_time_split(data_set)

train_ds <-
 training(data_split) \%>\%
 as_ts_dataset(Temp ~ Temp + index(Date), n_timesteps = 20, h = 1)

}
