% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-basic-rnn.R
\name{basic_rnn_fit}
\alias{basic_rnn_fit}
\title{Basic RNN model for time series forecasting}
\usage{
basic_rnn_fit(x, y, key, index, categorical_features = NULL)
}
\arguments{
\item{x}{Input features}

\item{y}{Targets}

\item{categorical_features}{List of categorical feature columns from data.frame x}

\item{optim}{An optimizer with arguments

If using simple fix_xy, we do not any assumption about time and number of items.}
}
\description{
Basic RNN model for time series forecasting
}
