% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-categorical.R
\name{cat2idx}
\alias{cat2idx}
\title{Encode vector of categorical features}
\usage{
cat2idx(x)
}
\arguments{
\item{x}{(character, factor or numeric) Vector of arbitrary length}
}
\value{
A list containing two values:
\itemize{
\item transformed_x - vector of transformed categorical features
\item dict - dictionary with category mapping
}
}
\description{
Encode vector of categorical features
}
\examples{
data("gss_cat", package = "forcats")
encode_categorical(gss_cat$marital)

}
