% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_is_complete}
\alias{check_is_complete}
\title{Check if input data contains no NAs.
Otherwise, return error.}
\usage{
check_is_complete(data)
}
\description{
Check if input data contains no NAs.
Otherwise, return error.
}
