% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_is_new_data_complete}
\alias{check_is_new_data_complete}
\title{Check if new data has NAs in columns others than predicted outcome}
\usage{
check_is_new_data_complete(object, new_data)
}
\description{
Check if new data has NAs in columns others than predicted outcome
}
