% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_recursion}
\alias{check_recursion}
\title{Check, if recursion should be used in forecasting}
\usage{
check_recursion(object, new_data)
}
\description{
Check, if recursion should be used in forecasting
}
