% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-deep-factor.R
\name{deep_factor}
\alias{deep_factor}
\title{Deep Factor Model with RNN model}
\usage{
deep_factor(
  freq,
  h,
  num_hidden_global = 50,
  num_layers_global = 1,
  num_factors = 10,
  num_hidden_local = 5,
  num_layers_local = 1,
  cell_type = c("lstm", "gru"),
  optim = torch::optim_adam,
  context_length = h,
  num_parallel_samples = 100,
  cardinality,
  embedding_dimension = 10,
  distr_output
)
}
\arguments{
\item{freq}{Time Series frequency}

\item{h}{Forecast horizon}

\item{num_hidden_global}{Number of units per hidden layer for the global RNN model (default: 50).}

\item{num_layers_global}{Number of hidden layers for the global RNN model (default: 1)}

\item{num_factors}{Number of global factors (default: 10).}

\item{num_hidden_local}{Number of units per hidden layer for the local RNN model (default: 5).}

\item{num_layers_local}{Number of hidden layers for the global local model (default: 1).}

\item{cell_type}{Type of recurrent cells to use (available: ‘lstm’ or ‘gru’; default: ‘lstm’).}

\item{optim}{A \code{torch} optimizer (default: torch::optim_adam)}

\item{context_length}{Training length (default: context_length = h)}

\item{num_parallel_samples}{Number of evaluation samples per time series to increase parallelism during inference.
This is a model optimization that does not affect the accuracy (default: 100).}

\item{cardinality}{List consisting of the number of time series (default: list(\link{1}).}

\item{embedding_dimension}{Dimension of the embeddings for categorical features (the same dimension is used for all embeddings, default: 10).}

\item{distr_output}{Distribution to use to evaluate observations and sample predictions (default: StudentTOutput()).}
}
\description{
Deep Factor with RNN models is an instance of
}
\note{
\href{https://arxiv.org/abs/1905.12417}{Deep Factors for Forecasting} by Wang Y. et al.
}
