% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{dict_size}
\alias{dict_size}
\title{Return size of categorical variables in the data.frame}
\usage{
dict_size(data)
}
\arguments{
\item{data}{(\code{data.frame}) A data.frame containing categorical variables.
The function automatically finds categorical variables,
calling internally \link{is_categorical} function.}
}
\value{
Named logical vector
}
\description{
Return size of categorical variables in the data.frame
}
\examples{
glimpse(tiny_m5)
dict_size(tiny_m5)

# We can choose only the features we want - otherwise it automatically
# selects logical, factor, character or integer vectors

tiny_m5 \%>\%
  select(store_id, event_name_1) \%>\%
  dict_size()

}
