% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equal-date.R
\name{equal}
\alias{equal}
\alias{equal_years}
\alias{equal_months}
\title{Adds a phantom 29th of February to data.frame}
\usage{
equal_years(.data, .date_col, .fill = TRUE)

equal_months(.data, .date_col, .fill = TRUE)
}
\arguments{
\item{.data}{A data.frame-like object}

\item{fill}{If true, all the dates between
minimum and maximum date are added}
}
\description{
Adds a phantom 29th of February to data.frame
}
\examples{
pedestrian_per_day <-
 as_tibble(tsibble::pedestrian) \%>\%
 group_by(Sensor, Date) \%>\%
 summarise(Count = sum(Count, na.rm = TRUE))

head(pedestrian_per_day)

# Equal years
filled_dates <-
  pedestrian_per_day \%>\%
  equal_years(Date)

head(filled_dates)

filter(filled_dates, is.na(Date))

# Equal months
filled_dates <-
  pedestrian_per_day \%>\%
  equal_months(Date)

filter(filled_dates, is.na(Date))
}
