% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heuristics.R
\name{geometric_pyramid}
\alias{geometric_pyramid}
\title{Geometric pyramid rule}
\usage{
geometric_pyramid(input_size, next_layer_size)
}
\arguments{
\item{input_size}{(\code{integer}) Input size}

\item{next_layer_size}{(\code{integer}) Next layer size}
}
\description{
A simple heuristics to choose hidden layer size
}
\references{
\href{https://books.google.de/books/about/Practical_Neural_Network_Recipes_in_C++.html?id=7Ez_Pq0sp2EC&redir_esc=y}{Practical Neural Network Recipes in C++}
}
