% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training-helpers.R
\name{get_x}
\alias{get_x}
\title{batch <- list(x_num = "aaa", x_cat = "bbb", y = "c")
get_x(batch)}
\usage{
get_x(batch)
}
\description{
batch <- list(x_num = "aaa", x_cat = "bbb", y = "c")
get_x(batch)
}
