% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-categorical.R
\name{idx2cat}
\alias{idx2cat}
\title{Invert \code{cat2idx} transformation}
\usage{
idx2cat(x, dict)
}
\arguments{
\item{x}{A vector with interger values from 1 to n}

\item{dict}{A dictionary with categorical variable mapping}
}
\description{
Invert \code{cat2idx} transformation
}
\examples{
# Encode factor vector
data("gss_cat", package = "forcats")
output <- encode_categorical(gss_cat$marital)

transformed_x <- output[[1]]
dict <- output[[2]]

dict

decode_categorical(transformed_x, dict)

}
