% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{init_gate_bias}
\alias{init_gate_bias}
\title{Initialize gates to pass full information}
\usage{
init_gate_bias(rnn_layer)
}
\description{
x <- list(rnn_layer = nn_lstm(2, 20))
init_gate_bias(x$rnn_layer)
x$rnn_layer$parameters$bias_ih_l1
}
\details{
~LSTM.bias_ih_l\link{k} – (b_ii|b_if|b_ig|b_io), of shape (4\emph{hidden_size)
~LSTM.bias_hh_l\link{k} – (b_hi|b_hf|b_hg|b_ho), of shape (4}hidden_size)
}
