% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{is_categorical}
\alias{is_categorical}
\title{Check, if vector is categorical, i.e.
if is logical, factor, character or integer}
\usage{
is_categorical(x)
}
\arguments{
\item{x}{A vector of arbitrary type}
}
\value{
Logical value
}
\description{
Check, if vector is categorical, i.e.
if is logical, factor, character or integer
}
\examples{
is_categorical(c(TRUE, FALSE, TRUE, FALSE, FALSE, FALSE, TRUE))
is_categorical(1:10)
is_categorical((1:10) + 0.1)
is_categorical(as.factor(c("Ferrari", "Lamborghini", "Porsche", "McLaren", "Koenigsegg")))
is_categorical(c("Ferrari", "Lamborghini", "Porsche", "McLaren", "Koenigsegg"))

}
