% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlp-module.R
\name{model_mlp}
\alias{model_mlp}
\title{A configurable feed forward network (Multi-Layer Perceptron)
with embedding}
\usage{
model_mlp(..., horizon, output_size, embedding = NULL, activation = nnf_relu)
}
\description{
A configurable feed forward network (Multi-Layer Perceptron)
with embedding
}
\examples{
net <- model_mlp(4, 2, 1)
x <- as_tensor(iris[, 1:4])
net(x)

# With categorical features
library(recipes)
iris_prep <-
   recipe(iris) \%>\%
   step_integer(Species) \%>\%
   prep() \%>\%
   juice()

iris_prep <- mutate(iris_prep, Species = as.integer(Species))

x_num <- as_tensor(iris_prep[, 1:4])
x_cat <- as_tensor(dplyr::select(iris_prep, 5))

n_unique_values <- dict_size(iris_prep)

.init_layer_spec <-
   init_layer_spec(
     num_embeddings = n_unique_values,
     embedding_dim  = embedding_size_google(n_unique_values),
     numeric_in     = 4,
     numeric_out    = 2
   )

net <- model_mlp(.init_layer_spec, 2, 1)

net(x_num, x_cat)

}
