% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn-mlp.R
\name{nn_mlp}
\alias{nn_mlp}
\title{A shortcut to create a feed-forward block (MLP block)}
\usage{
nn_mlp(..., activation = nnf_relu)
}
\arguments{
\item{...}{(\code{nn_module}, \code{function} \code{integer}, \code{character})
An arbitrary number of arguments, than can be:
* \code{nn_module} - e.g. \code{\link[torch:nn_relu]{torch::nn_relu()}}
* \code{function}  - e.g. \code{\link[torch:nnf_relu]{torch::nnf_relu}}
* \code{character} - e.g. \code{selu}, which is converted to \code{nnf_selu}
* \code{integer}   -}

\item{activation}{Used if only integers are specified. By default: \code{nnf_relu}}
}
\description{
A shortcut to create a feed-forward block (MLP block)
}
\examples{
nn_mlp(10, 1)
nn_mlp(30, 10, 1)

# Simple forward pass
net <- nn_mlp(4, 2, 1)
x <- as_tensor(iris[, 1:4])
net(x)

# Simple forward pass with identity function
net <- nn_mlp(4, 2, 1, activation = function (x) x)
x <- as_tensor(iris[, 1:4])
net(x)

}
