% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-nn-nonlinear.R
\name{nn_nonlinear}
\alias{nn_nonlinear}
\title{Shortcut to create linear layer with nonlinear activation function}
\usage{
nn_nonlinear(in_features, out_features, bias = TRUE, activation = nn_relu())
}
\arguments{
\item{in_features}{(\code{integer}) size of each input sample}

\item{out_features}{(\code{integer}) size of each output sample}

\item{bias}{(\code{logical}) If set to \code{FALSE}, the layer will not learn an additive bias.
Default: \code{TRUE}}

\item{activation}{(\code{nn_module}) A nonlinear activation function (default: \code{\link[torch:nn_relu]{torch::nn_relu()}})}
}
\description{
Shortcut to create linear layer with nonlinear activation function
}
\examples{
net <- nn_nonlinear(10, 1)
x   <- torch_tensor(matrix(1, nrow = 2, ncol = 10))
net(x)

}
