% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{nnf_mae}
\alias{nnf_mae}
\title{Mean Absolute Scaled Error}
\usage{
nnf_mae(input, target)
}
\arguments{
\item{input}{tensor (N,*) where ** means, any number of additional dimensions}

\item{target}{tensor (N,*) , same shape as the input}
}
\description{
Mean Absolute Scaled Error
}
\details{
Computed according to the formula:
\deqn{MAE = \frac{1}{n}\displaystyle\sum_{t=1}^{n}\abs{target - input}}
}
\seealso{
\code{\link[yardstick]{mae}}
}
