% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{nnf_mape}
\alias{nnf_mape}
\title{Mean Absolute Percentage Error}
\usage{
nnf_mape(input, target)
}
\arguments{
\item{input}{tensor (N,*) where ** means, any number of additional dimensions}

\item{target}{tensor (N,*) , same shape as the input}
}
\description{
Mean Absolute Percentage Error
}
\details{
Computed according to the formula:
\deqn{MAPE = \frac{1}{n}\displaystyle\sum_{t=1}^{n} \abs{\frac{target - input}{target}}}
}
\seealso{
\code{\link[yardstick]{mape}}
}
