% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equal-date.R
\name{numeric_date}
\alias{numeric_date}
\title{Converts a date column to numeric-only format}
\usage{
numeric_date(.date)
}
\arguments{
\item{.date}{A Date vector}
}
\value{
A numeric vector with YYYYMMDD format
}
\description{
Converts a date column to numeric-only format
}
\examples{
numeric_date(as.Date("2020-12-06"))

}
