% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_forecast}
\alias{plot_forecast}
\title{Plot forecast vs ground truth}
\usage{
plot_forecast(
  data,
  forecast,
  outcome,
  index = NULL,
  interactive = FALSE,
  title = "Forecast vs actual values",
  ...
)
}
\arguments{
\item{interactive}{(\code{logical})}
}
\description{
Plot forecast vs ground truth
}
