% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-extensions.R
\name{plug}
\alias{plug}
\title{Additional verbs for parsnip
A parsnip extension to plug some additional
training monitoring etc}
\usage{
plug()
}
\description{
Additional verbs for parsnip
A parsnip extension to plug some additional
training monitoring etc
}
