% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-ts-dataset.R
\name{predictors_spec}
\alias{predictors_spec}
\title{Predictors specification
It facilitates to keep the same variables in all the specification list
and avoid typos}
\usage{
predictors_spec(x_num = NULL, x_cat = NULL)
}
\description{
Predictors specification
It facilitates to keep the same variables in all the specification list
and avoid typos
}
