% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-data.R
\name{prepare_dl}
\alias{prepare_dl}
\title{Prepare dataloders}
\usage{
prepare_dl(
  data,
  formula,
  index,
  timesteps,
  horizon,
  categorical = NULL,
  validation = NULL,
  sample_frac = 1,
  batch_size,
  shuffle,
  jump,
  parsed_formula = NULL,
  flatten = FALSE,
  ...
)
}
\arguments{
\item{data}{(\code{data.frame}) An input data.frame object with.
For now only \strong{single} data frames are handled with no categorical features.}

\item{formula}{(\code{formula}) A formula describing, how to use the data}

\item{index}{(\code{character}) The index column name.}

\item{timesteps}{(\code{integer}) The time series chunk length.}

\item{horizon}{(\code{integer}) Forecast horizon.}

\item{categorical}{(\code{character}) Categorical features.}

\item{validation}{(\code{data.frame} or \code{numeric}) Validation dataset or percent of TODO.}

\item{sample_frac}{(\code{numeric}) Sample a fraction of rows (default: 1, i.e.: all the rows).}

\item{batch_size}{(\code{integer}) Batch size.}

\item{shuffle}{(\code{logical}) A dataloader argument - shuffle rows or not?}

\item{jump}{(\code{integer}) Input window shift.}
}
\description{
Prepare dataloders
}
