% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-internal.R
\name{print_and_capture}
\alias{print_and_capture}
\title{https://stackoverflow.com/questions/26083625/how-do-you-include-data-frame-output-inside-warnings-and-errors}
\usage{
print_and_capture(x)
}
\description{
https://stackoverflow.com/questions/26083625/how-do-you-include-data-frame-output-inside-warnings-and-errors
}
