% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-tensors.R
\name{resolve_data}
\alias{resolve_data}
\title{Prepare tensors}
\usage{
resolve_data(x, key, index, categorical_features = NULL)
}
\arguments{
\item{categorical_features}{}
}
\description{
Prepare tensors
}
