% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp-rnn.R
\name{rnn_fit}
\alias{rnn_fit}
\title{Basic RNN model for time series forecasting}
\usage{
rnn_fit(
  formula,
  data,
  learn_rate,
  hidden_units,
  dropout,
  timesteps = 20,
  horizon = 1,
  layers = 1,
  optim = optim_adam(),
  batch_size = 1,
  epochs = 10,
  loss_fn = nnf_mse_loss
)
}
\arguments{
\item{optim}{An optimizer with arguments}

\item{x}{Input features}

\item{y}{Targets}

\item{categorical_features}{List of categorical feature columns from data.frame x}
}
\description{
Basic RNN model for time series forecasting
}
