% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch-utils.R
\name{rnn_output_size}
\alias{rnn_output_size}
\title{RNN output size}
\usage{
rnn_output_size(module)
}
\arguments{
\item{module}{(nn_module) A torch \code{nn_module}}
}
\description{
RNN output size
}
\examples{
gru_layer <- nn_gru(15, 3)
rnn_output_size(gru_layer)
}
