% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{scale_params}
\alias{scale_params}
\title{Accessor for \code{scale_params} values in a dataloader object}
\usage{
scale_params(dataloader, ...)
}
\description{
Accessor for \code{scale_params} values in a dataloader object
}
