% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{set_device}
\alias{set_device}
\title{Set model device.}
\usage{
set_device(object, device, ...)
}
\arguments{
\item{object}{An neural network object.}

\item{device}{(\code{character}) Selected device.}
}
\value{
Object of the same class with device set.
}
\description{
Set model device.
}
\examples{
rnn_net <-
  model_rnn(
    input_size  = 1,
    output_size = 1,
    hidden_size = 10
  ) \%>\%
 set_device("cpu")

rnn_net

}
