% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/span-time.R
\name{span_time}
\alias{span_time}
\title{Span a time range each}
\usage{
span_time(.data, .key, .date_var, .by = "day")
}
\arguments{
\item{.data}{A data.frame-like object}

\item{.key}{A key column(s)}

\item{.date_var}{A Date-like format column}

\item{.by}{}
}
\description{
Span a time range each
}
\seealso{
\code{timetk::pad_by_time}
}
