% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-categorical.R
\name{step_cat2idx}
\alias{step_cat2idx}
\title{Encode categorical variables as integer values from 1 to n (an index)}
\usage{
step_cat2idx(
  recipe,
  ...,
  role = NA,
  trained = FALSE,
  columns = FALSE,
  skip = FALSE,
  id = rand_id("cat2int")
)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the
sequence of operations for this recipe.}

\item{...}{One or more selector functions to choose which
variables are affected by the step. See \code{\link[=selections]{selections()}}
for more details. For the \code{tidy} method, these are not
currently used.}
}
\description{
Encode categorical variables as integer values from 1 to n (an index)
}
\concept{factors}
\concept{preprocessing}
\concept{variable_encodings}
\keyword{datagen}
