% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torchts-model.R
\name{torchts_model}
\alias{torchts_model}
\title{Torchts abstract model}
\usage{
torchts_model(
  class,
  net,
  index,
  key,
  outcomes,
  predictors,
  optim,
  timesteps,
  horizon,
  scale
)
}
\description{
Torchts abstract model
}
