% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\docType{data}
\name{torchts_palette}
\alias{torchts_palette}
\title{Picked from torchts logo}
\format{
An object of class \code{character} of length 5.
}
\usage{
torchts_palette
}
\description{
Picked from torchts logo
}
\keyword{datasets}
