% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-formula.R
\name{torchts_parse_formula}
\alias{torchts_parse_formula}
\title{Parse time series formula}
\usage{
torchts_parse_formula(formula, data)
}
\description{
bidir
}
\examples{
parsed   <- torchts_parse_formula(value ~ . + backward(cat(snap_CA)) + backward(sell_price) + categ(wday), experiment_data)
parsed_2 <- torchts_parse_formula(value ~ . + backward(cat(snap_CA, wday)) + backward(sell_price) + categ(wday), experiment_data)
View(parsed_2)
}
