% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weather_pl}
\alias{weather_pl}
\title{Weather data from Polish "poles of extreme temperatures" in 2001-2020}
\format{
\describe{
\item{station}{A place where weather data were measured}
\item{date}{Date}
\item{tmax_daily}{Maximum daily air temperatury \link{C}}
\item{tmin_daily}{Minimum daily air temperature \link{C}}
\item{tmin_soil}{Minimum near surface air temperature \link{C}}
\item{rr_daily}{Total daily preciptation \link{mm}}
\item{rr_type}{Precipitation type \link{S/W}}
\item{rr_daytime}{Total precipitation during day \link{mm}}
\item{rr_nightime}{Total precipitation during night \link{mm}}
\item{press_mean_daily}{Daily mean pressure at station level \link{hPa}}
}
}
\usage{
weather_pl
}
\description{
The data comes from IMGW (Institute of Meteorology and Water Management) and
was downloaded using the \code{climate} package. Two places have been chosen:
\itemize{
\item{TRN - Tarnów ("pole of warmth")}
\item{SWK - Suwałki ("pole of cold")}
}
A subset of columns has been selected and \code{date} column was added.
}
\examples{
# Head of weather_pl
head(weather_pl)
}
\keyword{datasets}
