% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static.R
\name{which_static}
\alias{which_static}
\title{Check, which variables are static}
\usage{
which_static(data, key, cols = NULL)
}
\description{
Check, which variables are static
}
\examples{
data <- tiny_m5 \%>\%
dplyr::select(store_id, item_id, state_id,
              weekday, wday, month, year)

}
