% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-tensor.R
\name{as_tensor}
\alias{as_tensor}
\alias{as_tensor.default}
\alias{as_tensor.ts}
\alias{as_tensor.mts}
\alias{as_tensor.tsibble}
\title{Convert an object to tensor}
\usage{
as_tensor(
  .data,
  ...,
  dtype = NULL,
  device = NULL,
  requires_grad = FALSE,
  pin_memory = FALSE
)

\method{as_tensor}{default}(
  .data,
  dtype = NULL,
  device = NULL,
  requires_grad = FALSE,
  pin_memory = FALSE
)

\method{as_tensor}{ts}(
  .data,
  by = NULL,
  dtype = NULL,
  device = NULL,
  requires_grad = FALSE,
  pin_memory = FALSE
)

\method{as_tensor}{mts}(
  .data,
  ...,
  dtype = NULL,
  device = NULL,
  requires_grad = FALSE,
  pin_memory = FALSE
)

\method{as_tensor}{tsibble}(
  .data,
  ...,
  dtype = NULL,
  device = NULL,
  requires_grad = FALSE,
  pin_memory = FALSE
)
}
\arguments{
\item{.data}{A data.frame-like object}

\item{...}{Column names to wrap the data.frame-like object}

\item{dtype}{A torch_dtype instance}

\item{device}{A device created with torch_device()}

\item{requires_grad}{If autograd should record operations on the returned tensor.}

\item{pin_memory}{If set, returned tensor would be allocated in the pinned memory.}
}
\description{
The function converts an a data.frame-like object to torch_tensor instance.
If no column names are specified (as "three dots"), the function
simply transforms the input data to \code{matrix} and then to \code{torch_tensor}.

The second scenario assumes that we need a \code{torch_tensor},
which has more than two dimensions and some columns contains
indicators, how this wrapping should be performed.

It's especially useful when transforming a data.frame containing
multiple (and, possibly, multivariate) time series.
When passing optional column names, function:
\itemize{
\item arranges a data.frame by the given columns
\item removes these columns from the data.frame
\item creates a n-dimensional tensor with the following shape
(n_distinct(column_name_1), n_distinct(column_name_2), ..., number of other columns)
}
}
\examples{
# Simple data.frame-to-torch_tensor transformation
as_tensor(mtcars)

# Transformation with column-wise data wrapping
library(dplyr)
library(torchts)

# ts class - default
air_passengers <- as_tensor(AirPassengers)
class(air_passengers)
dim(air_passengers)

# ts class using data frequency
air_passengers <-
  AirPassengers \%>\%
  as_tensor(frequency(.))
class(air_passengers)
dim(air_passengers)

# ts class using arbitrary frequency

euro_stock

}
