% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-ts-dataloder.R
\name{as_ts_dataloader}
\alias{as_ts_dataloader}
\title{Quick shortcut to create a torch dataloader based on the given dataset}
\usage{
as_ts_dataloader(
  data,
  formula,
  index = NULL,
  key = NULL,
  target = NULL,
  n_timesteps,
  batch_size,
  h = 1,
  sample_frac = 1,
  scale = TRUE
)
}
\arguments{
\item{data}{(data.frame)}

\item{formula}{A formula describing, how to use the data}

\item{index}{The index column}

\item{key}{The key column(s)}

\item{n_timesteps}{The time seris chunk length}

\item{batch_size}{Batch size}

\item{h}{Forecast horizon}

\item{sample_frac}{Sample a fraction of rows (default: 1, i.e.: all the rows)}
}
\description{
Quick shortcut to create a torch dataloader based on the given dataset
}
\examples{
library(rsample)

suwalki_temp <-
   weather_pl \%>\%
   filter(station == "SWK") \%>\%
   select(date, temp = tmax_daily)

# Splitting on training and test
data_split <- initial_time_split(suwalki_temp)


}
