% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-ts-dataset.R
\name{as_ts_dataset}
\alias{as_ts_dataset}
\title{Create a torch dataset for time series data}
\usage{
as_ts_dataset(
  data,
  formula,
  index = NULL,
  key = NULL,
  target = NULL,
  timesteps,
  h = 1,
  sample_frac = 1,
  scale = TRUE
)
}
\arguments{
\item{data}{(data.frame)}

\item{formula}{A formula describing, how to use the data}

\item{index}{The index column}

\item{key}{The key column(s)}

\item{timesteps}{The time seris chunk length}

\item{h}{Forecast horizon}

\item{sample_frac}{Sample a fraction of rows (default: 1, i.e.: all the rows)}

\item{scale}{(logical) Scale feature columns}
}
\description{
Create a torch dataset for time series data
}
\note{
If \code{scale} is TRUE, only the input vaiables are scale and not the outcome ones.
See: \href{https://stats.stackexchange.com/questions/111467/is-it-necessary-to-scale-the-target-value-in-addition-to-scaling-features-for-re}{Is it necessary to scale the target value in addition to scaling features for regression analysis? (Cross Validated)}
}
\examples{
library(rsample)

suwalki_temp <-
   weather_pl \%>\%
   filter(station == "SWK") \%>\%
   select(date, temp = tmax_daily)

# Splitting on training and test
data_split <- initial_time_split(suwalki_temp)

train_ds <-
 training(data_split) \%>\%
 as_ts_dataset(temp ~ date, timesteps = 20, h = 1)

train_ds[1]

}
