% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-ts-dataset.R
\name{as_ts_dataset}
\alias{as_ts_dataset}
\title{Create a torch dataset for time series data from a \code{data.frame}-like object}
\usage{
as_ts_dataset(
  data,
  formula,
  index = NULL,
  key = NULL,
  predictors = NULL,
  outcomes = NULL,
  categorical = NULL,
  timesteps,
  horizon = 1,
  sample_frac = 1,
  scale = TRUE
)
}
\arguments{
\item{data}{(\code{data.frame}) An input data.frame object with.
For now only \strong{single} data frames are handled with no categorical features.}

\item{formula}{(\code{formula}) A formula describing, how to use the data}

\item{index}{(\code{character}) The index column name.}

\item{key}{(\code{character}) The key column name(s). Use only if formula was not specified.}

\item{predictors}{(\code{character}) Input variable names. Use only if formula was not specified.}

\item{outcomes}{(\code{character}) Target variable names. Use only if formula was not specified.}

\item{categorical}{(\code{character}) Categorical features.}

\item{timesteps}{(\code{integer}) The time series chunk length.}

\item{horizon}{(\code{integer}) Forecast horizon.}

\item{sample_frac}{(\code{numeric}) Sample a fraction of rows (default: 1, i.e.: all the rows).}

\item{scale}{(\code{logical} or \code{list}) Scale feature columns. Logical value or two-element list.
with values (mean, std)}
}
\description{
Create a torch dataset for time series data from a \code{data.frame}-like object
}
\note{
If \code{scale} is TRUE, only the input vaiables are scale and not the outcome ones.

See: \href{https://stats.stackexchange.com/questions/111467/is-it-necessary-to-scale-the-target-value-in-addition-to-scaling-features-for-re}{Is it necessary to scale the target value in addition to scaling features for regression analysis? (Cross Validated)}
}
\examples{
library(rsample)

suwalki_temp <-
   weather_pl \%>\%
   filter(station == "SWK") \%>\%
   select(date, temp = tmax_daily, rr_type)

# Splitting on training and test
data_split <- initial_time_split(suwalki_temp)

train_ds <-
 training(data_split) \%>\%
 as_ts_dataset(temp ~ date, timesteps = 20, horizon = 1)

train_ds[1]

}
