% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clear_outcome}
\alias{clear_outcome}
\title{Partially clear outcome variable
in new data by overriding with NA values}
\usage{
clear_outcome(data, index, outcome, timesteps, key = NULL)
}
\arguments{
\item{data}{(data.frame) New data}

\item{index}{Date variable}

\item{outcome}{Outcome (target) variable}

\item{timesteps}{Number of timesteps used by RNN model}

\item{key}{A key (id) to group the data.frame (for panel data)}
}
\description{
Partially clear outcome variable
in new data by overriding with NA values
}
