% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnn-module.R
\name{model_rnn}
\alias{model_rnn}
\title{A recurrent neural network model}
\usage{
model_rnn(layer = nn_gru, input_size, hidden_size, h, dropout = 0)
}
\arguments{
\item{layer}{A recurrent \code{torch} layer}

\item{input_size}{(integer) Input size}

\item{hidden_size}{(integer) Hidden layer size}

\item{h}{(integer) Horizon size}

\item{dropout}{(logical) Use dropout}
}
\description{
A recurrent neural network model
}
