% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{nnf_mape}
\alias{nnf_mape}
\title{Mean Absolute Percentage Error}
\usage{
nnf_mape(input, target)
}
\arguments{
\item{input}{tensor (N,*) where ** means, any number of additional dimensions}

\item{target}{tensor (N,*) , same shape as the input}
}
\description{
Mean Absolute Percentage Error
}
\details{
Computed according to the formula:
\deqn{MAPE = \frac{1}{n}\displaystyle\sum_{t=1}^{n} \left\|\frac{target - input}{target}\right\|}
}
\examples{
input <- c(92, 6.5, 57.69, 15.9, 88.47, 75.01, 5.06, 45.95, 27.8, 70.96)
input <- as_tensor(input)

target <- c(91.54, 5.87, 58.85, 10.73, 81.47, 75.39, 2.05, 40.95, 27.34, 66.61)
target <- as_tensor(target)

nnf_mape(input, target)

}
\seealso{
\link[yardstick:mape]{yardstick::mape}
}
