% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-rnn.R
\name{rnn}
\alias{rnn}
\title{General interface to recurrent neural netowrk models}
\usage{
rnn(
  mode = "regression",
  learn_rate = 0.01,
  epochs = 50,
  hidden_units = NULL,
  dropout = NULL,
  batch_size = 32
)
}
\arguments{
\item{mode}{(character) Model mode, default: 'regression'}

\item{learn_rate}{(numeric or dials::learn_rate) Learning rate}

\item{epochs}{(integer or dials::epochs) Number of epochs}

\item{hidden_units}{Number of hidden units}

\item{batch_size}{(integer) Batch size}
}
\description{
General interface to recurrent neural netowrk models
}
