% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnn-impl.R
\name{rnn_fit}
\alias{rnn_fit}
\title{Basic RNN model for time series forecasting}
\usage{
rnn_fit(
  formula,
  data,
  learn_rate,
  hidden_units,
  dropout,
  timesteps = 20,
  horizon = 1,
  optim = optim_adam(),
  validation = NULL,
  batch_size = 1,
  epochs = 10,
  scale = TRUE,
  loss_fn = nnf_mse_loss
)
}
\arguments{
\item{formula}{A formula describing, how to use the data}

\item{data}{(data.frame)}

\item{learn_rate}{(numeric) Learning rate}

\item{hidden_units}{(integer) Number of hidden units}
}
\description{
Basic RNN model for time series forecasting
}
