% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-formula.R
\name{torchts_parse_formula}
\alias{torchts_parse_formula}
\title{Parse time series formula}
\usage{
torchts_parse_formula(formula, data)
}
\description{
First version returns 3 types of variables:
\itemize{
\item outcome
\item predictor
\item index
}
}
\examples{
tarnow_temp <-
  weather_pl \%>\%
  filter(station == "TARNÓW") \%>\%
  select(date, temp = tmax_daily)

View(torchts_parse_formula(temp ~ temp +index(date), tarnow_temp))
View(torchts_parse_formula(temp ~ date), tarnow_temp))

# This example is not working
View(torchts_parse_formula(max_temp + min_temp ~ max_temp + min_temp + index(date), tarnow_temp))
TODO:
Rename
.type -> .role
.class -> .type
}
