% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-dataset.R
\name{ts_dataset}
\alias{ts_dataset}
\title{Create a time series dataset object}
\usage{
ts_dataset(
  .data,
  n_timesteps,
  h,
  input_columns = list(x = NULL),
  target_columns = list(y = NULL),
  sample_frac = 1
)
}
\arguments{
\item{.data}{(torch_tensor) An input data object}

\item{n_timesteps}{(integer) Number of timesteps for input tensor}

\item{h}{(integer) Forecast horizon: number of timesteps for output tensor}

\item{input_columns}{(list) Output specification}

\item{target_columns}{(list) Output specification}

\item{sample_fram}{(numeric) A numeric value > 0. and <= 1 to sample a subset of data}
}
\description{
Create a time series dataset object
}
\examples{
suppressMessages(library(dplyr))
library(torchts)
data("mts-examples", package="MTS")

ibmspko <-
    ibmspko \%>\%
    select(date, ibm)

ibm_tensor <- as_tensor(ibmspko, date)

ibm_dataset <-
    ts_dataset(ibm_tensor, n_timesteps = 7, h = 7)

ibm_dataset$.getitem(1)

}
