% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-dataset.R
\name{ts_dataset}
\alias{ts_dataset}
\title{Create a time series dataset object}
\usage{
ts_dataset(
  data,
  timesteps,
  h,
  jump = h,
  input_columns = list(x = NULL),
  target_columns = list(y = NULL),
  sample_frac = 1,
  scale = TRUE
)
}
\arguments{
\item{data}{(torch_tensor) An input data object}

\item{timesteps}{(integer) Number of timesteps for input tensor}

\item{h}{(integer) Forecast horizon: number of timesteps for output tensor}

\item{jump}{(integer) Jump length. By default: horizon length}

\item{input_columns}{(list) Output specification}

\item{target_columns}{(list) Output specification}

\item{scale}{(logical) Scale feature columns}

\item{sample_fram}{(numeric) A numeric value > 0. and <= 1 to sample a subset of data}
}
\description{
Create a time series dataset object
}
\note{
If \code{scale} is TRUE, only the input vaiables are scale and not the outcome ones.
See: \href{https://stats.stackexchange.com/questions/111467/is-it-necessary-to-scale-the-target-value-in-addition-to-scaling-features-for-re}{Is it necessary to scale the target value in addition to scaling features for regression analysis? (Cross Validated)}
}
\examples{
suppressMessages(library(dplyr))
library(torchts)
data("mts-examples", package="MTS")

ibmspko <-
    ibmspko \%>\%
    select(date, ibm)

ibm_tensor <- as_tensor(ibmspko, date)

ibm_dataset <-
    ts_dataset(ibm_tensor, timesteps = 7, h = 7)

ibm_dataset$.getitem(1)

}
